sched_yield if primary segment is known
Not a pretty hack, but a situation that is rather unique to us. If the
arcan backend is used in a session where the primary seg.  also comes
from arcan, we have a problem with aliasing and resize where the
underlying mapping might be resized while we are feeding buffers. To
work around that, there's mutex synchronization via the shmif_lock
approach, but that opens another problem. The mixer thread runs with
higher priority and the feed+synch might starve the lock in the other
thread. Until shmif can account for this problem, we have a hack where
the primary segment uses the userdata + magic cookie to establish a
resize flag to use for sched-yield.